requirements.txt

selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
loguru==0.7.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
schedule==1.2.0
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
